requirements.txt
# Core dependencies
crewai>=0.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.3.0
pydantic>=2.4.0

# Web server for Railway deployment
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0

# CrewAI tools
crewai-tools>=0.4.5

# Database for state management
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.0

# Utilities
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
pyjwt>=2.6.0
python-multipart>=0.0.6

# Supabase Integration
supabase>=2.3.0
postgrest>=0.13.0

# Background Tasks & Async Processing
celery>=5.3.0
redis>=5.0.0
aioredis>=2.0.1

# Social Media APIs
tweepy>=4.14.0
facebook-sdk>=3.1.0
python-linkedin-v2>=2.0.1
google-api-python-client>=2.100.0
youtube-api>=2.0.0

# Analytics & Business Intelligence
google-analytics-reporting>=0.7.0
google-ads>=21.0.0
pytrends>=4.9.0
semrush-python>=1.0.0

# Web Scraping & Data Collection
beautifulsoup4>=4.12.0
scrapy>=2.11.0
selenium>=4.15.0
playwright>=1.40.0

# Data Processing & Analysis
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Image & Video Processing
pillow>=10.0.0
opencv-python>=4.8.0
imageio>=2.31.0

# Caching & Performance
redis>=5.0.0
memcached>=1.59

# Rate Limiting & Security
slowapi>=0.1.9
python-limiter>=3.5.0
cryptography>=41.0.0
bcrypt>=4.0.0

# Monitoring & Logging
prometheus-client>=0.17.0
structlog>=23.2.0
sentry-sdk[fastapi]>=1.32.0

# Arabic Language Support
pyarabic>=0.6.15
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

# Email & Notifications
sendgrid>=6.10.0
twilio>=8.8.0
slack-sdk>=3.23.0

# File Storage & Processing
boto3>=1.29.0
azure-storage-blob>=12.19.0
google-cloud-storage>=2.10.0

# Webhook & Automation Integration
webhooks>=1.1.0
zapier-platform-core>=15.0.0
httpx>=0.25.0

# Configuration & Environment Management
dynaconf>=3.2.0
click>=8.1.0

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0

# Production Deployment
gunicorn>=21.2.0
gevent>=23.9.0
